requirements.txt
requests==2.32.5
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
gunicorn==23.0.0
requests==2.32.5
sqlparse==0.5.1
tzdata==2025.1 
django-cors-headers==4.9.0
